requirements.txt
streamlit
pandas
matplotlib
scipy
Pillow
pydicom
opencv-python
numpy
datetime
scikit-image
plotly
seaborn
simplejson
scikit-learn
requests